+++ src/shared/types/checkReactTypeSpec.js
@@ -11,10 +11,7 @@
 
 'use strict';
 
-var ReactPropTypesSecret = require('ReactPropTypesSecret');
-
-var invariant = require('invariant');
-var warning = require('warning');
+var checkPropTypes = require('checkPropTypes');
 
 var ReactComponentTreeHook;
 
@@ -31,20 +28,6 @@ if (
   ReactComponentTreeHook = require('ReactComponentTreeHook');
 }
 
-var loggedTypeFailures = {};
-
-/**
- * Assert that the values match with the type specs.
- * Error messages are memorized and will only be shown once.
- *
- * @param {object} typeSpecs Map of name to a ReactPropType
- * @param {object} values Runtime values that need to be type-checked
- * @param {string} location e.g. "prop", "context", "child context"
- * @param {string} componentName Name of the component for error messages.
- * @param {?object} element The React element that is being type-checked
- * @param {?number} workInProgressOrDebugID The React component instance that is being type-checked
- * @private
- */
 function checkReactTypeSpec(
   typeSpecs,
   values,
@@ -54,80 +37,35 @@ function checkReactTypeSpec(
   // It is only safe to pass fiber if it is the work-in-progress version, and
   // only during reconciliation (begin and complete phase).
   workInProgressOrDebugID,
-  // Default behavior is to suppress repeated warnings, but can be overridden
   warnOnRepeat
 ) {
-  for (var typeSpecName in typeSpecs) {
-    if (typeSpecs.hasOwnProperty(typeSpecName)) {
-      var error;
-      // Prop type validation may throw. In case they do, we don't want to
-      // fail the render phase where it didn't fail before. So we log it.
-      // After these have been cleaned up, we'll let them throw.
-      try {
-        // This is intentionally an invariant that gets caught. It's the same
-        // behavior as without this statement except with a better message.
-        invariant(
-          typeof typeSpecs[typeSpecName] === 'function',
-          '%s: %s type `%s` is invalid; it must be a function, usually from ' +
-          'React.PropTypes.',
-          componentName || 'React class',
-          location,
-          typeSpecName
-        );
-        error = typeSpecs[typeSpecName](values, typeSpecName, componentName, location, null, ReactPropTypesSecret);
-      } catch (ex) {
-        error = ex;
+  function formatMessage(message) {
+    if (__DEV__) {
+      let componentStackInfo = '';
+      if (!ReactComponentTreeHook) {
+        ReactComponentTreeHook = require('ReactComponentTreeHook');
       }
-      warning(
-        !error || error instanceof Error,
-        '%s: type specification of %s `%s` is invalid; the type checker ' +
-        'function must return `null` or an `Error` but returned a %s. ' +
-        'You may have forgotten to pass an argument to the type checker ' +
-        'creator (arrayOf, instanceOf, objectOf, oneOf, oneOfType, and ' +
-        'shape all require an argument).',
-        componentName || 'React class',
-        location,
-        typeSpecName,
-        typeof error
-      );
-      if (error instanceof Error && (warnOnRepeat || !(error.message in loggedTypeFailures))) {
-        // Only monitor this failure once because there tends to be a lot of the
-        // same error.
-        loggedTypeFailures[error.message] = true;
-
-        var componentStackInfo = '';
-
-        if (__DEV__) {
-          if (!ReactComponentTreeHook) {
-            ReactComponentTreeHook = require('ReactComponentTreeHook');
-          }
-          if (workInProgressOrDebugID != null) {
-            if (typeof workInProgressOrDebugID === 'number') {
-              // DebugID from Stack.
-              const debugID = workInProgressOrDebugID;
-              componentStackInfo = ReactComponentTreeHook.getStackAddendumByID(debugID);
-            } else if (typeof workInProgressOrDebugID.tag === 'number') {
-              // This is a Fiber.
-              // The stack will only be correct if this is a work in progress
-              // version and we're calling it during reconciliation.
-              const workInProgress = workInProgressOrDebugID;
-              componentStackInfo = ReactComponentTreeHook.getStackAddendumByWorkInProgressFiber(workInProgress);
-            }
-          } else if (element !== null) {
-            componentStackInfo = ReactComponentTreeHook.getCurrentStackAddendum(element);
-          }
+      if (workInProgressOrDebugID != null) {
+        if (typeof workInProgressOrDebugID === 'number') {
+          // DebugID from Stack.
+          const debugID = workInProgressOrDebugID;
+          componentStackInfo = ReactComponentTreeHook.getStackAddendumByID(debugID);
+        } else if (typeof workInProgressOrDebugID.tag === 'number') {
+          // This is a Fiber.
+          // The stack will only be correct if this is a work in progress
+          // version and we're calling it during reconciliation.
+          const workInProgress = workInProgressOrDebugID;
+          componentStackInfo = ReactComponentTreeHook.getStackAddendumByWorkInProgressFiber(workInProgress);
         }
-
-        warning(
-          false,
-          'Failed %s type: %s%s',
-          location,
-          error.message,
-          componentStackInfo
-        );
+      } else if (element !== null) {
+        componentStackInfo = ReactComponentTreeHook.getCurrentStackAddendum(element);
       }
+      message += componentStackInfo;
     }
+    return message;
   }
+
+  checkPropTypes(typeSpecs, values, location, componentName, formatMessage, warnOnRepeat);
 }
 
 module.exports = checkReactTypeSpec;
